[TASK] Update to the new Bootstrap::init() in TYPO3 v9.2
With https://review.typo3.org/c/55773/ the applications
no longer perform bootstrapping, therefore the entry
point scripts need to be adapted.

We can drop the custom CommandApplication as the entry
point leven is now passed as parameter to Bootstrap:init
from the the enty point script.

As the entry point change is breaking for releases <9.2,
this commit marks a new major version. (2.0)