[BUGFIX] Use correct autoload path, in case this package is linked
In current Composer versions, the binaries are not links any more
but files to allow the packages to be linked (e.g. from path repository).
For that Composer populates $GLOBALS['_composer_autoload_path']
in the created binary file to expose the correct path to the autoload
file for this installation.

So we should use this path if a current Composer version is used
and can gracefully fall back to directory traversal for older versions.